chore: project organization and refactoring
- extracted components to their own files
- created init files for pages and components
