--- models/registration.py
@@ -1,2 +1,0 @@
-from playwright.async_api import async_playwright
-
