+++ pages/__init__.py
@@ -1,0 +1,5 @@
+from .automation_portal import AutomationPortal
+from .checkout_page import CheckoutPage
+from .components import Navbar, CartSidebar, LoginPopup
+
+__all__ = ["AutomationPortal", "CheckoutPage", "Navbar", "CartSidebar", "LoginPopup"]
