+++ pages/checkout_page.py
@@ -1,8 +1,8 @@
-class Checkout:
-    def __init__(self, page):
-        self.page = page
+from playwright.async_api import Page
 
-        # selectors
+class CheckoutPage:
+    def __init__(self, page: Page):
+        self.page = page
         self.first_name = "#first-name"
         self.last_name = "#last-name"
         self.country = "#country"
@@ -23,6 +23,7 @@ class Checkout:
         await self.page.locator(selector).scroll_into_view_if_needed()
         await self.page.fill(selector, value)
 
+    # change this so it receives a list/dictionary maybe?
     async def fill_billing_details(
             self,
             first_name=None,
