+++ pages/components/__init__.py
@@ -1,0 +1,5 @@
+from .navbar import Navbar
+from .cart_sidebar import CartSidebar
+from .login_popup import LoginPopup
+
+__all__ = ["Navbar", "CartSidebar", "LoginPopup"]
